+++ components/AuthNav.js
@@ -1,9 +1,7 @@
 import React from 'react'
-import { connect } from 'react-redux'
 import { createStackNavigator } from '@react-navigation/stack'
 
 import { Login, Signup } from '../screens/AuthScreens'
-import { TabNavigator } from './AppNav'
 import WelcomeScreen from '../screens/WelcomeScreen'
 
 import colors from '../config/colors'
@@ -21,20 +19,6 @@ export const LoginSignup = () => (
     <Stack.Screen name="Signup" component={Signup} options={headerStyle} />
   </Stack.Navigator>
 )
-
-export const AppScreen = ({ user }) => {
-  return (
-    <React.Fragment>
-      {user && user.id ? <TabNavigator /> : <LoginSignup />}
-    </React.Fragment>
-  )
-}
-
-const mapState = (state) => ({
-  user: state.user,
-})
-
-export default connect(mapState, null)(AppScreen)
 
 const headerStyle = {
   headerStyle: { backgroundColor: colors.medium },
