+++ redux/recipe.js
@@ -1,5 +1,5 @@
 import axios from 'axios'
-import { URL } from './url'
+import { URL } from './serverUrl'
 
 // this entire reducer should only be used for managing state on RecipePostForm for a single recipe object. When the form is submitted, the thunk creator postRecipe is used to actually post that single recipe object to the DB.
