+++ screens/UserProfileScreen.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { SafeAreaView } from 'react-native'
+import { ScrollView } from 'react-native-gesture-handler'
 
 import UserProfileInfo from '../components/UserProfileInfo'
 import UserProfileRecipes from '../components/UserProfileRecipes'
@@ -9,8 +10,10 @@ import defaultStyles from '../config/defaultStyles'
 const UserProfileScreen = ({ navigation }) => {
   return (
     <SafeAreaView style={defaultStyles.container}>
-      <UserProfileInfo nav={navigation} />
-      <UserProfileRecipes nav={navigation} />
+      <ScrollView>
+        <UserProfileInfo nav={navigation} />
+        <UserProfileRecipes nav={navigation} />
+      </ScrollView>
     </SafeAreaView>
   )
 }
